move config, lifecycle, context to separate files
motivation: easier code navigation

changes: move config, lifecycle, context to separate source files
